Agentic-RAG-Pipeline/requirements.txt
google-generativeai>=0.7.2,<1.0.0
faiss-cpu>=1.8.0.post1,<2.0.0
httpx>=0.27.0,<1.0.0
requests>=2.32.0,<3.0.0
beautifulsoup4>=4.12.2,<5.0.0
pydantic>=2.7.0,<3.0.0
python-dotenv>=1.0.1,<2.0.0
numpy>=1.26.4,<3.0.0
